reorganizing files and deleting unnecessary files
