+++ src/app/layout.tsx
@@ -1,10 +1,14 @@
 import type { Metadata } from "next";
 import React from "react";
-import { Inter } from "next/font/google";
+import { Inter, Ceviche_One, Comic_Neue } from "next/font/google";
 import 'bootstrap/dist/css/bootstrap.css';
-import "./globals.css";
-import BootstrapClient from "./components/BootstrapClient";
+import "../styles/globals.css";
+import BootstrapClient from "../scripts/BootstrapClient";
 
+import Navbar from "../components/NavBar"
+import Footer from "../components/footer"
+
+//declaring const variables
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
@@ -20,7 +24,9 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={inter.className}>
+        <Navbar/>
         <BootstrapClient/>
+        <Footer/>
         {children}
       </body>
     </html>
